Rename monster-data.ts to monsterData.ts
